src/requirements.txt
altgraph==0.16.1
construct==2.9.45
cycler==0.10.0
future==0.16.0
glucometerutils==1
hidapi==0.7.99.post21
kiwisolver==1.0.1
macholib==1.10
matplotlib==2.2.2
numpy==1.13.3
pandas==0.23.3
pefile==2017.11.5
Pillow==5.2.0
pkg-resources==0.0.0
PyInstaller==3.3.1
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.7.3
pytz==2018.5
pyusb==1.0.2
reportlab==3.5.2
six==1.11.0
